+++ src/main/java/com/dsmpear/main/user_backend_v2/service/notice/NoticeService.java
@@ -6,5 +6,5 @@ import org.springframework.data.domain.Pageable;
 
 public interface NoticeService {
     NoticesResponse getNoticeList(Pageable page);
-    NoticeContentResponse getNoticeContent(Integer noticeId);
+    NoticeContentResponse getNoticeContent(Long noticeId);
 }
